+++ amethyst_renderer/src/pass/debug_lines/interleaved.rs
@@ -70,7 +70,7 @@ where
     V: Query<(Position, Color, Normal)>,
 {
     type Data = (
-        Option<Read<'a, ActiveCamera>>,
+        Read<'a, ActiveCamera>,
         ReadStorage<'a, Camera>,
         ReadStorage<'a, GlobalTransform>,
         WriteStorage<'a, DebugLinesComponent>, // DebugLines components