+++ amethyst_renderer/src/pass/pbm/separate.rs
@@ -79,7 +79,7 @@ impl DrawPbmSeparate {
 
 impl<'a> PassData<'a> for DrawPbmSeparate {
     type Data = (
-        Option<Read<'a, ActiveCamera>>,
+        Read<'a, ActiveCamera>,
         ReadStorage<'a, Camera>,
         Read<'a, AmbientColor>,
         Read<'a, AssetStorage<Mesh>>,