+++ amethyst_renderer/src/pass/shaded/separate.rs
@@ -78,7 +78,7 @@ impl DrawShadedSeparate {
 
 impl<'a> PassData<'a> for DrawShadedSeparate {
     type Data = (
-        Option<Read<'a, ActiveCamera>>,
+        Read<'a, ActiveCamera>,
         ReadStorage<'a, Camera>,
         Read<'a, AmbientColor>,
         Read<'a, AssetStorage<Mesh>>,