+++ amethyst_renderer/src/pass/skybox/interleaved.rs
@@ -45,7 +45,7 @@ impl DrawSkybox {
 
 impl<'a> PassData<'a> for DrawSkybox {
     type Data = (
-        Option<Read<'a, ActiveCamera>>,
+        Read<'a, ActiveCamera>,
         ReadStorage<'a, Camera>,
         ReadStorage<'a, GlobalTransform>,
         Read<'a, SkyboxColor>,